Nada funciona

+++ src/main/java/fatec/pweb/dao/ProdutoDAO.java
@@ -3,6 +3,6 @@ package fatec.pweb.dao;
 import fatec.pweb.dao.generic.DAOImpl;
 import fatec.pweb.model.Produto;
 
-public class ProdutoDAO extends DAOImpl<Produto, Long>{
+public class ProdutoDAO extends DAOImpl<Produto, Integer>{
 
 }
